+++ custom_components/eltako/sensor.py
@@ -344,7 +344,7 @@ async def async_setup_entry(
                         entities.append(EltakoMeterSensor(platform, gateway, dev_conf.id, dev_name, dev_conf.eep, SENSOR_DESC_WATER_CUMULATIVE, tariff=(tariff - 1)))
                         entities.append(EltakoMeterSensor(platform, gateway, dev_conf.id, dev_name, dev_conf.eep, SENSOR_DESC_WATER_CURRENT, tariff=(tariff - 1)))
 
-                elif dev_conf.eep in [A5_04_02, A5_10_12, A5_04_01]:
+                elif dev_conf.eep in [A5_04_01, A5_04_02, A5_04_03, A5_10_12]:
                     
                     entities.append(EltakoTemperatureSensor(platform, gateway, dev_conf.id, dev_name, dev_conf.eep))
                     entities.append(EltakoHumiditySensor(platform, gateway, dev_conf.id, dev_name, dev_conf.eep))
@@ -371,6 +371,11 @@ async def async_setup_entry(
                     entities.append(EltakoBatteryVoltageSensor(platform, gateway, dev_conf.id, dev_name, dev_conf.eep))
                     # _pir_status => as binary sensor
 
+                elif dev_conf.eep in [A5_06_01]:
+                    entities.append(EltakoIlluminationSensor(platform, gateway, dev_conf.id, dev_name, dev_conf.eep))
+                    #TODO: add twilight
+                    #TODO: add daylight
+                    # both are currently combined in illumination
 
             except Exception as e:
                 LOGGER.warning("[%s] Could not load configuration", platform)